pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "softice-mcp"
version = "0.1.0"
description = "MCP server bridge for SoftICE over a VT100 serial PTY"
readme = "README.md"
requires-python = ">=3.12"
authors = [{ name = "Gabriel Speckhahn" }]
dependencies = ["pyte>=0.8"]

[project.scripts]
softice-mcp = "softice_mcp.server:main"

[tool.setuptools]
packages = ["softice_mcp"]
